change project license to Apache License Version 2.0
This change of license shall allow the extensions to be used in broader
base of projects, especially commercial ones, that otherwise could not
profit from the toolkit nor contribute back to it.

I, as the owner of the code, agree to the change of license and are
more than happy to contribute the SAP CX Tools to the community.